+++ IOSMatchingEngineSDK/MobiledgeXiOSLibrary/Classes/MatchingEngine/FindCloudlet.swift
@@ -139,7 +139,7 @@ extension MobiledgeXiOSLibrary.MatchingEngine {
     ///   - port: port override of the dme server port
     ///   - request: FindCloudlet dictionary, from createFindCloudletReqwuest.
     /// - Returns: API Dictionary/json
-    public func findCloudlet(host: String, port: UInt, request: [String: Any])
+    public func findCloudlet(host: String, port: UInt16, request: [String: Any])
         -> Promise<[String: AnyObject]>
     {
         os_log("Finding nearest Cloudlet appInsts matching this MatchingEngine client.", log: OSLog.default, type: .debug)
